Add a `target-tag` input to control the tag we produce.
Currently, we always target `:latest` and produce an ambiguous date-based tag.

With this change, `:latest` is the default and the date-based tags will be prefixed with `:latest-`, but this behavior can be overridden to target `:foo` and `:foo-` prefixed instead or in addition.

My hope with this is to better support producing multiple flavors of images from a single repo, e.g. a `:root` image from `distroless/git` to ease the Tekton transition (building from something like `.apko-root.yaml`.

Signed-off-by: Matt Moore <mattmoor@chainguard.dev>