Refactor image-refs outputs
This changes apko-publish to emit the image-refs output, and for
apko-snapshot to pass it through.

The image_refs input is still accepted for compatibility, but
DEPRECATED, with consistent defaults to /tmp/apko.images

The output is a JSON-formatted array of strings.

After this change we'll need to update all the consumers of these
actions accordingly, then remove the image_refs input.

Signed-off-by: Jason Hall <jason@chainguard.dev>