This stops passing "now" as the `SOURCE_DATE_EPOCH` to `apko`.
We have made changes to `apko` to compute the "build date epoch" based on `MAX(installed packages builddate)`, which should make the image timestamp a reflection of the last time one of the packages contributing to the image changed.

This should result in significantly less nightly churn from image rebuilds where only the timestamp changes.

Signed-off-by: Matt Moore <mattmoor@chainguard.dev>